Update DSA_P120. DÃY CON DÀI NHẤT.cpp

+++ D21 - THI CHINH THUC - DEMO/CA 16H/DSA_P120. DÃY CON DÀI NHẤT.cpp
@@ -3,7 +3,7 @@ using namespace std;
 int main(){
     int n,m;
     cin >> n >> m;
-    long long a[n+5]; ll b[m+5];
+    long long a[n+5]; long long b[m+5];
     int dp[n+5][m+5];
     for(int i=1;i<=n;i++) cin >> a[i];
     for(int i=1;i<=m;i++) cin >> b[i];
